release candidate 0 - version 15
Restructured sources into layers: search (engine), atoms (geometry, cell, formats), common (utilities), xtalopt (application)
Removed Eigen, OpenBabel, GenXrdPattern download and dependency; Avogadro RPC and backtrace debug helpers
Added new structure view and xrd view widgets to GUI
Removed the generic optimizer (now, only gulp, vasp, pwscf, castep, siesta, mtp)
Added new libraries to external: xraylib, libmsym, args
GUI fully split as a build option
Added a centralized compatibility layer for platform and Qt/Qwt
Moved all settings/satae I/O to xtalopt (engine no longer do or participate in that)
Support for older formats collected in xtalopt/legacy; update state version to 5
Unified all xtalopt settings into a registry table (their accessors, defaults, properties) with a single validation
Revised command-line options
Added a structure toolset to the xtalopt for structural analysis and molecular unit/crystal generation
Revised queue interface (direct vs batch run, etc) and SSH usage for remoteQueue
Separated objectives into objectives (min/max) and constraints (fil)
Added/revised settings flags (sshMethod,directRunCommand,forcedSpgs,usingScaledIADs, pareto filter zero weight, debug output)
Unified input settings syntax (potcar/psf = <id> <file>) and defined asset file group for optimizers
Improved search wall time for fast local optimizers (queue manager workflow, background file saving, in-memory parent-selection data)
Expanded and updated the test suite
Updated supporting data (schemes, scripts, tests/sampledata) for format consistency with new version
Build system: support for Qt6/Qt5, system SSH always with libssh optional; macOS code-sign; CPack on all platforms, 3rd party licenses

+++ tests/sshconnection_clitest.cpp
@@ -1,6 +1,5 @@
 /**********************************************************************
-  SSHConnection - SSHConnectionTest class provides unit testing for the
-  SSHConnection class
+  SSHConnection - Unit test for the SSHConnection class: cli-ssh
 
   Copyright (C) 2010-2011 David C. Lonie
 
@@ -11,41 +10,47 @@
   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
   See the License for the specific language governing permissions and
   limitations under the License.
- **********************************************************************/
+ ***********************************************************************/
 
-#include <globalsearch/sshconnection_cli.h>
+#include <search/ssh/sshconnection_cli.h>
+
+#include "sshtestconfig.h"
+
+#include <common/fileutils.h>
 
 #include <QString>
+#include <QTemporaryDir>
 #include <QTemporaryFile>
 #include <QtTest>
 
-using namespace GlobalSearch;
+#include <algorithm>
+
+using namespace Search;
 
 class SSHConnectionCLITest : public QObject
 {
   Q_OBJECT
 
 private:
-  SSHConnectionCLI* conn;
+  SSHConnectionCLI* conn = nullptr;
+  bool m_testFilesCreated = false;
+  QTemporaryDir m_localTempRoot;
   QTemporaryFile m_localTempFile;
   QString m_remoteFileName;
   QString m_localNewFileName;
   QString m_fileContents;
 
   // Create a local directory structure:
-  // [tmp path]/sshtesttmp/
-  //                       testfile1
-  //                       newdir/
-  //                              testfile2
+  // [temporary path]/source/
+  //                         testfile1
+  //                         newdir/
+  //                                testfile2
   QStringList m_dirLayout;
   QDir m_localTempDir;
   QString m_remoteDir;
   QString m_localNewDir;
   QString m_testfile1Contents;
   QString m_testfile2Contents;
-
-  // Large file test
-  QTemporaryFile m_largeTempFile;
 
 private slots:
   /**
@@ -69,14 +74,7 @@ private slots:
   void cleanup();
 
   // Tests
-  //  void isValid();
-  //  void isConnected1();
-  //  void disconnectSession();
-  //  void isConnected2();
-  //  void reconnectSession();
-
   void execute();
-  //  void executeLargeOutput();
 
   void copyFileToServer();
   void readRemoteFile();
@@ -87,12 +85,14 @@ private slots:
   void readRemoteDirectoryContents();
   void copyDirectoryFromServer();
   void removeRemoteDirectory();
-
-  void largeFileCopyBenchmark();
 };
 
 void SSHConnectionCLITest::initTestCase()
 {
+  const SSHTestConfig config = readSSHTestConfig(false);
+  if (!config.enabled)
+    QSKIP(qPrintable(config.skipMessage));
+
   // Write local file for later manipulation
   m_fileContents = "This is a test file.\n\nIt has text in it.";
   m_localTempFile.open();
@@ -105,54 +105,46 @@ void SSHConnectionCLITest::initTestCase()
   m_localNewFileName = m_localTempFile.fileName() + ".new";
 
   // Create a local directory structure:
-  // [tmp path]/sshtesttmp/
-  //                       testfile1
-  //                       newdir/
-  //                              testfile2
+  // [temporary path]/source/
+  //                         testfile1
+  //                         newdir/
+  //                                testfile2
   m_remoteDir = ".sshtmpdir";
-  m_dirLayout << m_remoteDir + "/testfile1" << m_remoteDir + "/newdir/"
-              << m_remoteDir + "/newdir/testfile2";
-  m_localTempDir.mkpath(QDir::tempPath() + "/sshtesttmp");
-  m_localTempDir.mkpath(QDir::tempPath() + "/sshtesttmp/newdir");
-  m_localTempDir.setPath(QDir::tempPath() + "/sshtesttmp");
-  m_localNewDir = m_localTempDir.path() + ".new";
+  const QString remoteNewDir = Common::remotePath(m_remoteDir, "newdir");
+  m_dirLayout << Common::remotePath(m_remoteDir, "testfile1")
+              << remoteNewDir + "/"
+              << Common::remotePath(remoteNewDir, "testfile2");
+  QVERIFY(m_localTempRoot.isValid());
+  const QString localTestDir = Common::localPath(m_localTempRoot.path(), "source");
+  const QString localNewSubdir = Common::localPath(localTestDir, "newdir");
+  m_localTempDir.mkpath(localTestDir);
+  m_localTempDir.mkpath(localNewSubdir);
+  m_localTempDir.setPath(localTestDir);
+  m_localNewDir = Common::localPath(m_localTempRoot.path(), "copy");
 
-  QFile testfile1(m_localTempDir.path() + "/testfile1");
+  QFile testfile1(Common::localPath(m_localTempDir.path(), "testfile1"));
   testfile1.open(QIODevice::WriteOnly);
   QTextStream teststream1(&testfile1);
   m_testfile1Contents = "This is the first file's contents.\n";
   teststream1 << m_testfile1Contents;
   testfile1.close();
 
-  QFile testfile2(m_localTempDir.path() + "/newdir/testfile2");
+  QFile testfile2(Common::localPath(localNewSubdir, "testfile2"));
   testfile2.open(QIODevice::WriteOnly);
   QTextStream teststream2(&testfile2);
   m_testfile2Contents = "and these are the second's.\n";
   teststream2 << m_testfile2Contents;
   testfile2.close();
 
-  // Create a large file, 10 MB
-  // / Open local file
-  m_largeTempFile.open();
-  QTextStream lts(&m_largeTempFile);
-  // / Create buffer
-  QString buffer(1048576, '0');
-  // / Write
-  for (int i = 0; i < 10; i++)
-    lts << buffer;
-  m_largeTempFile.close();
+  m_testFilesCreated = true;
 
   // Open ssh connection
   try {
-    // Ensure that the following points to a real server, acct, and pw
-    // combo. Do not commit any changes here! (considering using
-    // /etc/hosts to map "testserver" to a real server with a
-    // chroot-jailed acct/pw = "test")
     conn = new SSHConnectionCLI();
-    conn->setLoginDetails("testserver", "test", "test");
+    conn->setLoginDetails(config.host, config.user, config.password, config.port);
     //    conn->connectSession();
   } catch (SSHConnection::SSHConnectionException) {
-    conn = 0;
+    conn = nullptr;
     QFAIL("Cannot connect to ssh server. Make sure that the connection opened "
           "in initTestCase() points to a valid account on a real host before "
           "debugging this failure.");
@@ -161,23 +153,24 @@ void SSHConnectionCLITest::initTestCase()
 
 void SSHConnectionCLITest::cleanupTestCase()
 {
-  QFile::remove(m_localTempFile.fileName());
-  QFile::remove(m_localNewFileName);
+  if (m_testFilesCreated) {
+    QFile::remove(m_localTempFile.fileName());
+    QFile::remove(m_localNewFileName);
 
-  QFile::remove(m_localTempDir.path() + "/testfile1");
-  QFile::remove(m_localTempDir.path() + "/newdir/testfile2");
-  m_localTempDir.rmdir(m_localTempDir.path() + "/newdir");
-  m_localTempDir.rmdir(m_localTempDir.path());
-  QFile::remove(m_localNewDir + "/testfile1");
-  QFile::remove(m_localNewDir + "/newdir/testfile2");
-  m_localTempDir.rmdir(m_localNewDir + "/newdir");
-  m_localTempDir.rmdir(m_localNewDir);
-
-  QFile::remove(m_largeTempFile.fileName());
+    QFile::remove(Common::localPath(m_localTempDir.path(), "testfile1"));
+    QFile::remove(Common::localPath(
+      Common::localPath(m_localTempDir.path(), "newdir"), "testfile2"));
+    m_localTempDir.rmdir(Common::localPath(m_localTempDir.path(), "newdir"));
+    m_localTempDir.rmdir(m_localTempDir.path());
+    QFile::remove(Common::localPath(m_localNewDir, "testfile1"));
+    QFile::remove(Common::localPath(Common::localPath(m_localNewDir, "newdir"), "testfile2"));
+    m_localTempDir.rmdir(Common::localPath(m_localNewDir, "newdir"));
+    m_localTempDir.rmdir(m_localNewDir);
+  }
 
   if (conn)
     delete conn;
-  conn = 0;
+  conn = nullptr;
 }
 
 void SSHConnectionCLITest::init()
@@ -194,24 +187,18 @@ void SSHConnectionCLITest::execute()
   QString stdout_str, stderr_str;
   int ec;
 
-  // Execute the command 10 times.
-  QBENCHMARK_ONCE
-  {
-    for (int i = 1; i <= 10; i++) {
-      QVERIFY2(conn->execute(command, stdout_str, stderr_str, ec),
-               QString("Execution of \'" + command + "\' (#" +
-                       QString::number(i) + ") failed.")
-                 .toStdString()
-                 .c_str());
-      QCOMPARE(ec, 0);
-      QCOMPARE(stdout_str, QString("6\n"));
-      QVERIFY2(stderr_str.isEmpty(),
-               QString("Execution of \'" + command + "\' produced an error: " +
-                       stderr_str)
-                 .toStdString()
-                 .c_str());
-    }
-  }
+  // Only one execution to verify the command channel works.
+  QVERIFY2(conn->execute(command, stdout_str, stderr_str, ec),
+           QString("Execution of \'" + command + "\' failed.")
+             .toStdString()
+             .c_str());
+  QCOMPARE(ec, 0);
+  QCOMPARE(stdout_str, QString("6\n"));
+  QVERIFY2(stderr_str.isEmpty(),
+           QString("Execution of \'" + command + "\' produced an error: " +
+                   stderr_str)
+             .toStdString()
+             .c_str());
 }
 
 void SSHConnectionCLITest::copyFileToServer()
@@ -252,8 +239,9 @@ void SSHConnectionCLITest::copyDirectoryToServer()
   QVERIFY2(conn->copyDirectoryToServer(m_localTempDir.path(), m_remoteDir),
            "Error copying directory to server.");
   QString cont1, cont2;
-  conn->readRemoteFile(m_remoteDir + "/testfile1", cont1);
-  conn->readRemoteFile(m_remoteDir + "/newdir/testfile2", cont2);
+  conn->readRemoteFile(Common::remotePath(m_remoteDir, "testfile1"), cont1);
+  conn->readRemoteFile(Common::remotePath(Common::remotePath(m_remoteDir, "newdir"), "testfile2"),
+                       cont2);
   QCOMPARE(cont1, m_testfile1Contents);
   QCOMPARE(cont2, m_testfile2Contents);
 }
@@ -262,8 +250,8 @@ void SSHConnectionCLITest::readRemoteDirectoryContents()
 {
   QStringList contents;
   QVERIFY(conn->readRemoteDirectoryContents(m_remoteDir, contents));
-  qSort(contents);
-  qSort(m_dirLayout);
+  std::sort(contents.begin(), contents.end());
+  std::sort(m_dirLayout.begin(), m_dirLayout.end());
   QCOMPARE(contents, m_dirLayout);
 }
 
@@ -272,12 +260,12 @@ void SSHConnectionCLITest::copyDirectoryFromServer()
   QVERIFY2(conn->copyDirectoryFromServer(m_remoteDir, m_localNewDir),
            "Error copying directory from server.");
 
-  QFile f1(m_localNewDir + "/testfile1");
+  QFile f1(Common::localPath(m_localNewDir, "testfile1"));
   f1.open(QIODevice::ReadOnly);
   QString cont1(f1.readAll());
   QCOMPARE(cont1, m_testfile1Contents);
 
-  QFile f2(m_localNewDir + "/newdir/testfile2");
+  QFile f2(Common::localPath(Common::localPath(m_localNewDir, "newdir"), "testfile2"));
   f2.open(QIODevice::ReadOnly);
   QString cont2(f2.readAll());
   QCOMPARE(cont2, m_testfile2Contents);
@@ -287,26 +275,6 @@ void SSHConnectionCLITest::removeRemoteDirectory()
 {
   QVERIFY2(conn->removeRemoteDirectory(m_remoteDir),
            "Error removing remote directory.");
-}
-
-void SSHConnectionCLITest::largeFileCopyBenchmark()
-{
-  QBENCHMARK
-  {
-    conn->copyFileToServer(m_largeTempFile.fileName(), ".sshlargetest");
-  }
-  QBENCHMARK
-  {
-    conn->copyFileFromServer(".sshlargetest",
-                             m_largeTempFile.fileName() + ".new");
-  }
-  conn->removeRemoteFile(".sshlargetest");
-  QFile largeCopy(m_largeTempFile.fileName() + ".new");
-  largeCopy.open(QIODevice::ReadOnly);
-  m_largeTempFile.open();
-  QCOMPARE(largeCopy.readAll(), m_largeTempFile.readAll());
-  largeCopy.close();
-  m_largeTempFile.close();
 }
 
 QTEST_MAIN(SSHConnectionCLITest)
